+++ src/components/index.ts
@@ -32,6 +32,7 @@ export * from './FrontifyPattern';
 export * from './Grid';
 export * from './InlineDialog';
 export * from './InputLabel';
+export * from './LegacyOrderableList';
 export * from './LinkChooser';
 export * from './LoadingBar';
 export * from './LoadingCircle';
